+++ webapi/Lib/PersistenceHelpers/MccSoft.PersistenceHelpers/DomainEvents/DomainEventsTransactionInterceptor.cs
@@ -3,7 +3,7 @@ using System.Data.Common;
 using System.Threading;
 using System.Threading.Tasks;
 using MccSoft.DomainHelpers.DomainEvents;
-using MediatR;
+using Mediator;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.Logging;